backend/pyproject.toml
[tool.poetry]
name = "social-app"
version = "0.1.0"
description = ""
authors = ["mrtedn21 <bezgin.sasha06@gmail.com>"]
readme = "README.md"
packages = [{include = "social_app"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.11.0"
psycopg2-binary = "^2.9.3"
drf-yasg = "^1.21.3"
django-filter = "^22.1"
certifi = "^2022.6.15"
celery = "^5.2.7"
redis = "^4.3.4"
django-cors-headers = "^3.13.0"
django-ckeditor = "^6.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
